Module fast3d::decode

Fast3D command decoding.

This module provides two methods for decoding commands:

decode_f3d_command

decode_f3d_display_list

Both of these transform RawF3DCommands into F3DCommands. Since some commands have multiple parts, the former returns a DecodeResult, which may require additional raw commands to complete.

Note: this module is not complete and may have errors.

Structs

BlendMode

ColorCombineMode

CombineMode

F3DCommandIter

F3DWrapMode

FillColor

Either rgba5551 or zdz (z = 14 bits, dz = 2 bits)

GeometryModes

Image

PrimDepth

RawF3DCommand

A raw Fast3D command for decoding.

Rectangle

RenderMode

RenderModeFlags

Rgba32

TextureBlock

TextureRectangle

TileIndex

TileParams

TileSize

Unimplemented

Enums

AlphaCompare

AlphaDither

BlendAlpha1

BlendAlpha2

BlendColor

ColorCombineComponent

ColorDither

ComponentSize

CvgDst

CycleType

DPCommand

DecodeResult

The result of decoding a raw command.

DepthSource

F3DCommand

A decoded Fast3D command.

ImageFormat

MatrixMode

MatrixOp

PipelineMode

SPCommand

ScissorMode

TextureConvert

TextureDetail

TextureFilter

TextureLUT

ZMode

Functions

decode_f3d_command

Decodes a raw Fast3D command.

decode_f3d_display_list

Decodes a stream of RawF3DCommands into a stream of F3DCommands.