fast3d
Module interpret

A display list interpreter that produces a self-contained F3DRenderData object that is straightforward to render.

After implementing F3DMemory for loading vertex/texture/dls/etc data from memory, interpret_f3d_display_list can be called.

Note: this module is not complete and may have errors.

Structs

ColorExpr

An expression for the fragment color output.

DrawCommand

One triangle list draw call.

F3DRenderData

An object containing processed render data from a display list.

PipelineId

An id for the pipeline state of a draw call.

PipelineInfo

The pipeline state used for a draw call.

SamplerState

Sampling parameters for a texture.

ScreenRectangle

A rectangle in screen space.

TextureData

Rgba8 texture buffer.

TextureIndex

An id for a loaded texture.

TextureState

Texture pixel data and sampling parameters.

Enums

ColorArg

An argument to ColorExpr.

CullMode

WrapMode

Traits

F3DMemory

A trait with methods for reading from game memory.

Functions

interpret_f3d_display_list

Processes memory.root_dl() and returns draw data in a simpler to render and self-contained F3DRenderData object.