Trait fast3d::interpret::F3DMemory

pub trait F3DMemory {
    type Ptr: Debug + Copy + PartialEq;
    type Error: From<F3DError>;
    type DlIter<'a>: Iterator<Item = Result<F3DCommand<Self::Ptr>, Self::Error>>
       where Self: 'a;

    // Required methods
    fn root_dl(&self) -> Result<Self::DlIter<'_>, Self::Error>;
    fn read_dl(&self, ptr: Self::Ptr) -> Result<Self::DlIter<'_>, Self::Error>;
    fn read_u8(
        &self,
        dst: &mut [u8],
        ptr: Self::Ptr,
        offset: usize
    ) -> Result<(), Self::Error>;
    fn read_u16(
        &self,
        dst: &mut [u16],
        ptr: Self::Ptr,
        offset: usize
    ) -> Result<(), Self::Error>;
    fn read_u32(
        &self,
        dst: &mut [u32],
        ptr: Self::Ptr,
        offset: usize
    ) -> Result<(), Self::Error>;

    // Provided methods
    fn read_vertices_default(
        &self,
        dst: &mut [Vertex],
        ptr: Self::Ptr,
        offset: usize
    ) -> Result<(), Self::Error> { ... }
    fn read_vertices(
        &self,
        dst: &mut [Vertex],
        ptr: Self::Ptr,
        offset: usize
    ) -> Result<(), Self::Error> { ... }
}

A trait with methods for reading from game memory.

This needs to be implemented so that interpret_f3d_display_list can read objects from memory (viewports, textures, display lists, etc).

Required Associated Types

type Ptr: Debug + Copy + PartialEq

The pointer type that can be read from a display list command.

PartialEq is only used for checking if the color image is equal to the depth image.

type Error: From<F3DError>

Error that can be thrown when reading from memory.

type DlIter<'a>: Iterator<Item = Result<F3DCommand<Self::Ptr>, Self::Error>> where Self: 'a

An iterator over a display list that is read from memory.

Required Methods

fn root_dl(&self) -> Result<Self::DlIter<'_>, Self::Error>

Returns the top level display list to be interpreted.

fn read_dl(&self, ptr: Self::Ptr) -> Result<Self::DlIter<'_>, Self::Error>

Reads a child display list from memory at the given address.

fn read_u8( &self, dst: &mut [u8], ptr: Self::Ptr, offset: usize ) -> Result<(), Self::Error>

Reads dst.len() u8s from memory, starting at ptr + offset (in bytes).

fn read_u16( &self, dst: &mut [u16], ptr: Self::Ptr, offset: usize ) -> Result<(), Self::Error>

Reads dst.len() u16s from memory, starting at ptr + offset (in bytes).

fn read_u32( &self, dst: &mut [u32], ptr: Self::Ptr, offset: usize ) -> Result<(), Self::Error>

Reads dst.len() u32s from memory, starting at ptr + offset (in bytes).

Provided Methods

fn read_vertices_default( &self, dst: &mut [Vertex], ptr: Self::Ptr, offset: usize ) -> Result<(), Self::Error>

Reads dst.len() vertices from memory, starting at ptr + offset (in bytes).

Callers should use read_vertices instead.

fn read_vertices( &self, dst: &mut [Vertex], ptr: Self::Ptr, offset: usize ) -> Result<(), Self::Error>

Reads dst.len() vertices from memory, starting at ptr + offset (in bytes).