Module fast3d::decode

Fast3D command decoding.

This module provides two methods for decoding commands:

decode_f3d_command

decode_f3d_display_list

Both of these transform RawF3DCommands into F3DCommands. Since some commands have multiple parts, the former returns a DecodeResult, which may require additional raw commands to complete.

Currently these functions panic on invalid commands.

Note: this module is not complete and may have errors.

Structs

F3DCommandIter

RawF3DCommand

A raw Fast3D command for decoding.

Enums

DecodeResult

The result of decoding a raw command.

Functions

decode_f3d_command

Decodes a raw Fast3D command.

decode_f3d_display_list

Decodes a stream of RawF3DCommands into a stream of F3DCommands.